+++ CHANGELOG.md
@@ -19,6 +19,7 @@
   - On flyouts, the comparison uses the **actual items** in your inventory (e.g., evaluates both trinket slots), not just the equipped slot
 
 - **Health Macro:** Option to use **combat potions** in combat
+- **Health Macro:** Custom priority order to allow either _custom spells_, _potions_, _healthstones_ or _combat potions_ to be used
 
 - **GameMenu size:** Option to scale the Game Menu